fix test on windows

+++ graphblas/tests/test_matrix.py
@@ -4090,7 +4090,7 @@ def test_to_from_edgelist(A):
     result = Matrix.from_edgelist([[0, 1, 10], [2, 3, 20]])
     expected = Matrix.from_coo([0, 2], [1, 3], [10, 20])
     assert expected.isequal(result, check_dtype=True)
-    result = Matrix.from_edgelist(np.array([[0, 1, 10], [2, 3, 20]]))
+    result = Matrix.from_edgelist(np.array([[0, 1, 10], [2, 3, 20]], dtype=np.int64))
     assert expected.isequal(result, check_dtype=True)
 
     result = Matrix.from_edgelist([], nrows=2, ncols=3, dtype=int)
